@@ -14,19 +14,17 @@
     "moment": "^2.24.0",
     "qrcode.react": "^0.9.3",
     "querystring": "^0.2.0",
-    "react": "^16.9.0",
-    "react-dom": "^16.9.0",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
     "react-hot-loader": "^4.12.11",
     "react-iframe": "^1.8.0",
     "react-redux": "^7.1.1",
     "react-redux-loading-bar": "^4.4.0",
     "react-router-dom": "^5.0.1",
-    "react-scripts": "3.1.1",
+    "react-scripts": "4.0.3",
     "react-toastify": "^5.3.2",
-    "react-video-js-player": "^1.1.1",
     "reactstrap": "^8.0.1",
     "redux": "^4.0.4",
-    "redux-reset": "^0.3.0",
     "redux-thunk": "^2.3.0",
     "typescript": "^3.6.2",
     "video.js": "^7.15.4",
@@ -36,8 +34,7 @@
     "lint": "eslint -c .eslintrc.js --ext .js,.jsx,.ts,.tsx .",
     "lint:fix": "npm run lint -- --fix",
     "start": "REACT_APP_STAGE=dev react-scripts start",
-    "build": "REACT_APP_STAGE=production ./before_build.sh && react-app-rewired build",
-    "pub": "REACT_APP_STAGE=production react-app-rewired build",
+    "build": "bash ./before_build.sh && react-app-rewired build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
     "postbuild": "bash ./post_build_osx",
