@@ -20,8 +20,9 @@ module.exports = {
   ],
   plugins: ['react', 'react-hooks'],
   rules: {
-    // Place to specify ESLint rules. Can be used to overwrite rules specified from the extended configs
-    // e.g. "@typescript-eslint/explicit-function-return-type": "off",
+    // Place to specify ESLint rules. Can be used to overwrite rules specified from
+    // the extended configs e.g. "@typescript-eslint/explicit-function-return-type":
+    // "off",
     'no-use-before-define': 'off',
     '@typescript-eslint/no-use-before-define': ['error'],
     'react/prop-types': 0,
