@@ -1,23 +1,20 @@
 import React from 'react'
 import TableContainer from '../../Components/Tables/TableContainer'
-import TagCreate from '../../Tags/create'
+import TagCreate from '../../tags/create'
 import resources from '../locale'
-import {Link} from 'react-router-dom';
+import { Link } from 'react-router-dom'
 
 export default class GroupCreate extends React.Component {
-  
-
   render() {
     return (
       <TableContainer title={resources.group}>
-      <div className="mb-4">
-            <Link to={`/device`} className="btn btn-default">
-              返回
-            </Link>{" "}
-          </div>
-        <TagCreate {...this.props}/>
+        <div className="mb-4">
+          <Link to={`/device`} className="btn btn-default">
+            返回
+          </Link>{' '}
+        </div>
+        <TagCreate {...this.props} />
       </TableContainer>
     )
   }
-
 }
