@@ -3,26 +3,6 @@ import { uuidv4 } from "../lib/util";
 import { webosApis } from "lgservice";
 const qs = require("querystring");
 
-export const requestDeviceInfo = async() => {
-    const { queryDeviceInfo, queryosInfo } = webosApis.systemservice;
-    const deviceInfo = await queryDeviceInfo();
-    const osInfo = await queryosInfo();
-    const { wired_addr, bt_addr, wifi_addr, returnValue, device_name } = deviceInfo;
-    const {
-        webos_name, //webOS,OSE
-        webos_release
-    } = osInfo;
-    return {
-        wired_addr, 
-        bt_addr, //Bluetooth address.
-        wifi_addr,
-        returnValue,
-        device_name,
-        webos_name,
-        webos_release
-    };
-};
-
 export const requestQR = () => {
     var url = `${process.env.REACT_APP_AUTH_URL}/api/authSession`;
     var sessionId = uuidv4().replace(/-/g, "");
@@ -41,6 +21,25 @@ export const requestToken = sessionId => {
     return axios({ url, method: "get" });
 };
 
+export const requestDeviceInfo = async() => {
+    const { queryDeviceInfo, queryosInfo } = webosApis.systemservice;
+    const deviceInfo = await queryDeviceInfo();
+    const osInfo = await queryosInfo();
+    const { wired_addr, bt_addr, wifi_addr, returnValue, device_name } = deviceInfo;
+    const {
+        webos_name, //webOS,OSE
+        webos_release
+    } = osInfo;
+    return {
+        wired_addr, bt_addr, //Bluetooth address.
+        wifi_addr,
+        returnValue,
+        device_name,
+        webos_name,
+        webos_release
+    };
+};
+
 export const postDeviceInfo = (token, authorizeCode, deviceInfo) => {
     var url = `${process.env.REACT_APP_MEMBER_URL}/api/License/UploadDeviceInfo`;
     const { wired_addr } = deviceInfo;
@@ -50,9 +49,10 @@ export const postDeviceInfo = (token, authorizeCode, deviceInfo) => {
         data: qs.stringify({
             deviceId: wired_addr,
             authorizeCode,
-            ...{
-                deviceInfo
-            }
+            os: deviceInfo.webos_name,
+            os_ver: deviceInfo.webos_release,
+            mac: wired_addr,
+            name: deviceInfo.device_name
         }),
         headers: {
             "Content-Type": "application/x-www-form-urlencoded",
