@@ -12,13 +12,17 @@ import {
 import { useDispatch, useSelector } from "react-redux";
 import { useHistory } from "react-router";
 
-export default() => {
-    const RunStep = Object.freeze({ RequestQR: 1, RequestToken: 2, RequestLicense: 3, Finished: 4 });
-    const { exists,mkdir } = webosApis.webosFileService;
-    const [runStep,
-        setRunStep] = useState(RunStep.RequestQR);
-    const [id,
-        setId] = useState();
+export default () => {
+    const RunStep = Object.freeze({
+        RequestQR: 1,
+        RequestToken: 2,
+        RequestLicense: 3,
+        Finished: 4
+    });
+    const { exists, mkdir } = webosApis.webosFileService;
+    const { autoboot, startServer } = webosApis.bootservice;
+    const [runStep, setRunStep] = useState(RunStep.RequestQR);
+    const [id, setId] = useState();
     const delay = 5000;
     const dispatch = useDispatch();
     const history = useHistory();
@@ -27,6 +31,16 @@ export default() => {
     const { APP_ROOT } = config;
 
     useEffect(() => {
+        autoboot()
+            .then(() => {
+                return startServer();
+            })
+            .then(res => {
+                console.log("startserver", res);
+            })
+            .catch(e => {
+                console.log(e);
+            });
         dirReady();
         dispatch(requestQR());
     }, []);
@@ -39,16 +53,19 @@ export default() => {
     }, []);
 
     const dirReady = () => {
-        exists(APP_ROOT).then(exist => {
-            if (!exist) {
-                return mkdir(APP_ROOT);
-            }
-            return true;
-        }).then((res) => {
-            console.log(res);
-        }).catch(e => {
-            console.log("mkdsDir", e);
-        });
+        exists(APP_ROOT)
+            .then(exist => {
+                if (!exist) {
+                    return mkdir(APP_ROOT);
+                }
+                return true;
+            })
+            .then(res => {
+                console.log(res);
+            })
+            .catch(e => {
+                console.log("mkdsDir", e);
+            });
     };
 
     const stateLoop = () => {
@@ -58,12 +75,13 @@ export default() => {
         } else if (!token.access_token && runStep === RunStep.RequestToken) {
             dispatch(requestToken(QR.authorizeCode));
         } else if (token.access_token && runStep === RunStep.RequestToken) {
-            dispatch(requestInstance(token.access_token));
             setRunStep(RunStep.RequestLicense);
             requestDeviceInfo().then(deviceInfo => {
                 const { wired_addr } = deviceInfo;
                 setId(wired_addr);
-                dispatch(postDeviceInfo(token.access_token, deviceInfo, QR.authorizeCode));
+                dispatch(
+                    postDeviceInfo(token.access_token, QR.authorizeCode, deviceInfo)
+                );
             });
         } else if (!license.resourceServer && runStep === RunStep.RequestLicense) {
             dispatch(requestLicense(token.access_token, id));
@@ -74,17 +92,14 @@ export default() => {
         }
     };
 
-    const saveLicense = (token) => {
+    const saveLicense = token => {
         license.token = token.access_token;
         console.log("save license", license);
-        config
-            .instance
-            .licenseWrite(license)
-            .then(err => {
-                if (err) {
-                    console.log("write license file error", err);
-                }
-            });
+        config.instance.write(license).then(err => {
+            if (err) {
+                console.log("write license file error", err);
+            }
+        });
     };
 
     const divStyles = {
@@ -97,11 +112,15 @@ export default() => {
         <div className="container-fluid" style={ divStyles }>
             <div className="row">
                 <div className="col">
-                    { QR.qrUrl && <QRCode value={ QR.qrUrl } size={ 256 } className="float-right"/> }
+                    { QR.qrUrl &&
+                        <QRCode value={ QR.qrUrl } size={ 256 } className="float-right" /> }
                 </div>
-                <div className="col" style={ {
-                    padding: "5%"
-                } }>
+                <div
+                    className="col"
+                    style={ {
+                        padding: "5%"
+                    } }
+                >
                     <h1>Scan QR code to activate device,Please.</h1>
                 </div>
             </div>
