@@ -6,7 +6,6 @@ import IndexReducer from "./index-reducer";
 //import { loadingBarMiddleware } from "react-redux-loading-bar";
 import { applyMiddleware, compose, createStore } from "redux";
 import thunk from "redux-thunk";
-import reduxReset from "redux-reset";
 import { Route, BrowserRouter as Router, Switch } from "react-router-dom";
 import config from "./config";
 // Polyfill, modifying the global Object
@@ -16,7 +15,7 @@ const composeSetup = process.env.NODE_ENV !== "production" && typeof window === 
   ? window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__
   : compose;
 /*eslint-enable */
-const store = createStore(IndexReducer, composeSetup(applyMiddleware(thunk), reduxReset()));
+const store = createStore(IndexReducer, composeSetup(applyMiddleware(thunk)));
 
 var render = Component => (
     <Provider store={ store }>
