@@ -6,7 +6,6 @@ using Ioliz.Service.Models;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
-using Newtonsoft.Json;
 
 namespace Ioliz.Service.Controllers
 {
@@ -24,7 +23,7 @@ namespace Ioliz.Service.Controllers
     }
 
     [HttpGet("/api/topic/{catelog}")]
-    public string KeyValues(TopicCatelog catelog){
+    public IActionResult KeyValues(TopicCatelog catelog){
       if(catelog == TopicCatelog.SensorModel || 
       catelog == TopicCatelog.PWMConfig ||
       catelog == TopicCatelog.BoardInfo ||
@@ -39,26 +38,26 @@ namespace Ioliz.Service.Controllers
     }
     //获取全局数据字典
    // [HttpGet("/api/topic/global/{catelog}")]
-     string GetGlobalTitles(TopicCatelog catelog)
+     IActionResult GetGlobalTitles(TopicCatelog catelog)
     {
       var single = ctx.Topics.FirstOrDefault(x => x.Catelog == catelog);
       if (single == null)
       {
-        return "";
+        return Json(new string[]{});
       }
-      return single.Content;
+      return Json(single.Content);
     }
 
     //获取用户键值列表
     //[HttpGet("/api/topic/user/{catelog}")]
-     string GetUserTitles(TopicCatelog catelog)
+     IActionResult GetUserTitles(TopicCatelog catelog)
     {
       var single = ctx.Topics.FirstOrDefault(x => x.Catelog == catelog && x.UserName == User.Identity.Name);
       if (single == null)
       {
-        return "";
+        return Json(new string[]{});
       }
-      return single.Content;
+      return Json(single.Content);
     
     }
 
